Fixing linux build.
<==== Mercurial data ====>
Branch: master

+++ mediaserver/src/rest/handlers/file_system_handler.cpp
@@ -36,7 +36,7 @@ int QnFileSystemHandler::executeGet(const QString &, const QnRequestParamList &p
     
     QnStorageStatusReply reply;
     reply.pluginExists = storage;
-    reply.storage.storageId = exists ? storage->getId() : -1;
+    reply.storage.storageId = exists ? storage->getId().toInt() : -1;
     reply.storage.path = storageUrl;
     reply.storage.freeSpace = storage ? storage->getFreeSpace() : -1;
     reply.storage.reservedSpace = storage ? storage->getSpaceLimit() : -1;
